Moved SSLContext generation/setting to helper class
